pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sysdig-cli"
version = "0.1.0-alpha"
description = "Sysdig Platform CLI - Alpha (built by Sergej Epp)"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
sysdig = "sysdig_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
sysdig_cli = ["assets/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 130

[tool.ruff.lint]
# C901: too complex (generated command dispatch code is intentionally complex)
# E402: module-level imports not at top (necessary for registration pattern in main.py)
extend-ignore = ["C901", "E402"]
